refactor: modify the project structure
